Add files via upload
A Hospital Management Web Mini Project is a simplified web-based system designed to streamline and automate the key functions of a hospital or clinic. It manages essential aspects such as about hospital, doctors, appointment scheduling, and services.  It also includes features for managing doctors, staff, and hospital inventory. This project enhances the efficiency of hospital operations, reduces paperwork, and improves patient care .